replaced DataContractSerializer with protobuf-net
