+++ PEExplorer/packages.config
@@ -5,6 +5,7 @@
   <package id="Microsoft.Diagnostics.Runtime.Latest" version="0.8.31.3" targetFramework="net462" />
   <package id="NLog" version="4.4.10" targetFramework="net462" />
   <package id="Prism.Core" version="6.3.0" targetFramework="net452" />
+  <package id="protobuf-net" version="2.2.1" targetFramework="net462" />
   <package id="SharpDisasm" version="1.1.5" targetFramework="net452" />
   <package id="Zodiacon.WPF" version="1.2.3" targetFramework="net462" />
 </packages>
